network: splitting VPN Gateway/VPN Connection & renaming to match the generated type name
