+++ azurerm/internal/services/network/vpn_gateway_connection_resource.go
@@ -33,7 +33,7 @@ func resourceArmVPNGatewayConnection() *schema.Resource {
 		Delete: resourceArmVpnGatewayConnectionResourceDelete,
 
 		Importer: azSchema.ValidateResourceIDPriorToImport(func(id string) error {
-			_, err := parse.VPNGatewayConnectionID(id)
+			_, err := parse.VpnConnectionID(id)
 			return err
 		}),
 
@@ -337,7 +337,7 @@ func resourceArmVpnGatewayConnectionResourceCreateUpdate(d *schema.ResourceData,
 		return fmt.Errorf("empty or nil ID returned for Vpn Gateway Connection Resource %q (Resource Group %q / VPN Gateway: %q) ID", name, gatewayId.ResourceGroup, gatewayId.Name)
 	}
 
-	id, err := parse.VPNGatewayConnectionID(*resp.ID)
+	id, err := parse.VpnConnectionID(*resp.ID)
 	if err != nil {
 		return err
 	}
@@ -352,25 +352,25 @@ func resourceArmVpnGatewayConnectionResourceRead(d *schema.ResourceData, meta in
 	ctx, cancel := timeouts.ForRead(meta.(*clients.Client).StopContext, d)
 	defer cancel()
 
-	id, err := parse.VPNGatewayConnectionID(d.Id())
+	id, err := parse.VpnConnectionID(d.Id())
 	if err != nil {
 		return err
 	}
 
-	resp, err := client.Get(ctx, id.ResourceGroup, id.Gateway, id.Name)
+	resp, err := client.Get(ctx, id.ResourceGroup, id.VpnGatewayName, id.Name)
 	if err != nil {
 		if utils.ResponseWasNotFound(resp.Response) {
-			log.Printf("[DEBUG] Vpn Gateway Connection Resource %q was not found in VPN Gateway %q in Resource Group %q - removing from state!", id.Name, id.Gateway, id.ResourceGroup)
+			log.Printf("[DEBUG] Vpn Gateway Connection Resource %q was not found in VPN Gateway %q in Resource Group %q - removing from state!", id.Name, id.VpnGatewayName, id.ResourceGroup)
 			d.SetId("")
 			return nil
 		}
 
-		return fmt.Errorf("retrieving Vpn Gateway Connection Resource %q (Resource Group %q / VPN Gateway %q): %+v", id.Name, id.ResourceGroup, id.Gateway, err)
+		return fmt.Errorf("retrieving Vpn Gateway Connection Resource %q (Resource Group %q / VPN Gateway %q): %+v", id.Name, id.ResourceGroup, id.VpnGatewayName, err)
 	}
 
 	d.Set("name", id.Name)
 
-	gatewayId := parse.NewVPNGatewayID(id.ResourceGroup, id.Gateway)
+	gatewayId := parse.NewVPNGatewayID(id.ResourceGroup, id.VpnGatewayName)
 	d.Set("vpn_gateway_id", gatewayId.ID(subscriptionId))
 
 	if prop := resp.VpnConnectionProperties; prop != nil {
@@ -409,24 +409,24 @@ func resourceArmVpnGatewayConnectionResourceDelete(d *schema.ResourceData, meta 
 	ctx, cancel := timeouts.ForDelete(meta.(*clients.Client).StopContext, d)
 	defer cancel()
 
-	id, err := parse.VPNGatewayConnectionID(d.Id())
+	id, err := parse.VpnConnectionID(d.Id())
 	if err != nil {
 		return err
 	}
 
-	locks.ByName(id.Gateway, VPNGatewayResourceName)
-	defer locks.UnlockByName(id.Gateway, VPNGatewayResourceName)
+	locks.ByName(id.VpnGatewayName, VPNGatewayResourceName)
+	defer locks.UnlockByName(id.VpnGatewayName, VPNGatewayResourceName)
 
-	future, err := client.Delete(ctx, id.ResourceGroup, id.Gateway, id.Name)
+	future, err := client.Delete(ctx, id.ResourceGroup, id.VpnGatewayName, id.Name)
 	if err != nil {
 		if response.WasNotFound(future.Response()) {
 			return nil
 		}
-		return fmt.Errorf("deleting Vpn Gateway Connection Resource %q (Resource Group %q / VPN Gateway %q): %+v", id.Name, id.ResourceGroup, id.Gateway, err)
+		return fmt.Errorf("deleting Vpn Gateway Connection Resource %q (Resource Group %q / VPN Gateway %q): %+v", id.Name, id.ResourceGroup, id.VpnGatewayName, err)
 	}
 	if err = future.WaitForCompletionRef(ctx, client.Client); err != nil {
 		if !response.WasNotFound(future.Response()) {
-			return fmt.Errorf("Error waiting for the deletion of VPN Gateway Connection %q (Resource Group %q / VPN Gateway %q): %+v", id.Name, id.ResourceGroup, id.Gateway, err)
+			return fmt.Errorf("Error waiting for the deletion of VPN Gateway Connection %q (Resource Group %q / VPN Gateway %q): %+v", id.Name, id.ResourceGroup, id.VpnGatewayName, err)
 		}
 	}
